+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+### Added
+- Add Certes.Acme.Resource.Directory to replace Certes.Acme.AcmeDirectory
 
 ## 1.0.7 - 2017-07-20
 ### Changed